实现微信注册登录功能

+++ src/main/java/com/easyorder/entity/Order.java
@@ -19,7 +19,7 @@ public class Order {
 	Long orderId;
 	String orderDesc;
 	Integer orderState;
-	Integer orderAmount;
+	Float orderAmount;
 	Date createTime;
 	Date payTime;
 	String orderEvaluation;//评价
